Tmp commit

+++ src/main/java/woody/core/relations/Relateable.java
@@ -16,14 +16,18 @@ import sirius.db.mixing.annotations.Transient;
 import sirius.kernel.di.std.Part;
 
 /**
- * Created by aha on 11.01.17.
+ * Embedded into an entity to make it eligible for incoming relations.
  */
 public class Relateable extends Composite {
 
     @Transient
     protected final Entity owner;
 
-
+    /**
+     * Creates and initializes the composite for the given entity.
+     *
+     * @param owner the entity which is the target for relations
+     */
     public Relateable(Entity owner) {
         this.owner = owner;
     }
@@ -34,10 +38,7 @@ public class Relateable extends Composite {
     @BeforeDelete
     protected void onDelete() {
         if (owner != null && !owner.isNew()) {
-            oma.select(Relation.class)
-               .eq(Relation.TARGET, owner.getUniqueName())
-               .delete();
+            oma.select(Relation.class).eq(Relation.TARGET, owner.getUniqueName()).delete();
         }
     }
-
 }
